We fix in multired2.v a problem similar to the one in multired.v.
By lack of time, we adopt however a lazier solution.

The problem is that lemma r''_0_eq, due to some call to omega using
one more variable of the context (in an irrelevant way), depends on
extra variable.

In multired.v, we tracked where omega was failing, but this was
costly. Here, we simply accept that the lemma depends useless on one
more variable and we instantiate this extra variable with an arbitrary
inhabitant.